+++ app/src/main/java/com/example/practice/domain/usecase/DeleteUserUseCase.kt
@@ -4,5 +4,5 @@ import com.example.practice.domain.models.UserModel
 import com.example.practice.domain.repository.DataBase
 
 class DeleteUserUseCase(val dataBase: DataBase) {
-    suspend fun deleteItem(id: Int) = dataBase.deleteItem(id)
+    suspend fun deleteItem(id: Long) = dataBase.deleteItem(id)
 }
